requirements.txt
psycopg2-binary==2.9.9
luigi==3.5.1
python-dotenv==1.0.1
dbt-postgres==1.8.2
